pyproject.toml
[project]
name = "fomina-shop-api"
version = "1.3.0"
description = "API for fomina-shop"
authors = [{ name = "VodoGamer", email = "vodogamer@yandex.ru" }]
dependencies = [
	"fastapi>=0.110.2",
	"pydantic>=2.7.0",
	"uvicorn>=0.29.0",
	"sqlalchemy>=2.0.29",
	"alembic>=1.13.1",
	"asyncpg>=0.29.0",
	"python-dotenv>=1.0.1",
	"sqladmin>=0.16.1",
	"fastapi-storages>=0.3.0",
	"pillow>=10.3.0",
	"itsdangerous>=2.2.0",
	"yookassa>=3.3.0",
	"telegrinder>=0.1.dev171",
	"aiohttp>=3.10.5",
	"orjson>=3.10.7",
]
readme = "README.md"
requires-python = ">= 3.12"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
lint.select = ["I", "E", "F", "B", "Q"]
lint.ignore = ["B008"]
line-length = 99

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/api"]

[dependency-groups]
dev = ["pre-commit>=4.0.1", "pyright>=1.1.390", "ruff>=0.8.3"]
